Major chages. Fully propagated Map object and new Go To algorithm. The Load and Save game no longer functions because there is so much more data and differences between save format.

+++ src/Graphics/CRectangle.cs
@@ -4,24 +4,24 @@ namespace OpenCiv1.Graphics
 {
 	public class CRectangle
 	{
-		public short ScreenID = 0;
-		public short Left = 0;
-		public short Top = 0;
-		public short Width = 320;
-		public short Height = 200;
+		public int ScreenID = 0;
+		public int Left = 0;
+		public int Top = 0;
+		public int Width = 320;
+		public int Height = 200;
 		public ushort Flags = 0;
 		public byte FrontColor = 0;
 		public byte PixelMode = 0;
 		public byte BackColor = 15;
 		public byte Undefined0 = 0;
-		public short FontID = 0;
+		public int FontID = 0;
 
 		public CRectangle()
 		{
 		}
 
-		public CRectangle(short screenID, short left, short top, short width, short height,
-			ushort flags, byte frontColor, byte pixelMode, byte backColor, byte undefined0, short fontID)
+		public CRectangle(int screenID, int left, int top, int width, int height,
+			ushort flags, byte frontColor, byte pixelMode, byte backColor, byte undefined0, int fontID)
 		{
 			this.ScreenID = screenID;
 			this.Left = left;
